src/Python/AI.Client/requirements.txt
openai
langgraph 
langchain_openai
ipykernel
azure-identity
mcp
python-dotenv
anthropic